Split protocol class to own file because was getting too big
